test: move 'make_task' / 'make_graph' utilities to conftest
