+++ test/test_optimize.py
@@ -13,6 +13,8 @@ from taskgraph.optimize.base import All, Any, Not, OptimizationStrategy
 from taskgraph.task import Task
 from taskgraph.taskgraph import TaskGraph
 
+from .conftest import make_graph, make_task
+
 
 class Remove(OptimizationStrategy):
     def should_remove_task(self, task, params, arg):
@@ -33,46 +35,6 @@ def default_strategies():
         "remove": Remove(),
         "replace": Replace(),
     }
-
-
-def make_task(
-    label,
-    optimization=None,
-    task_def=None,
-    optimized=None,
-    task_id=None,
-    dependencies=None,
-    if_dependencies=None,
-):
-    task_def = task_def or {
-        "sample": "task-def",
-        "deadline": {"relative-datestamp": "1 hour"},
-    }
-    task = Task(
-        kind="test",
-        label=label,
-        attributes={},
-        task=task_def,
-        if_dependencies=if_dependencies or [],
-    )
-    task.optimization = optimization
-    task.task_id = task_id
-    if dependencies is not None:
-        task.task["dependencies"] = sorted(dependencies)
-    return task
-
-
-def make_graph(*tasks_and_edges, **kwargs):
-    tasks = {t.label: t for t in tasks_and_edges if isinstance(t, Task)}
-    edges = {e for e in tasks_and_edges if not isinstance(e, Task)}
-    tg = TaskGraph(tasks, Graph(set(tasks), edges))
-
-    if kwargs.get("deps", True):
-        # set dependencies based on edges
-        for l, r, name in tg.graph.edges:
-            tg.tasks[l].dependencies[name] = r
-
-    return tg
 
 
 def make_opt_graph(*tasks_and_edges):
